fix: apply design reticle to pairing scanner via shared QRScannerFrame
