+++ app/src/bcsc-theme/features/account-transfer/transferee/TransferQRScannerScreen.test.tsx
@@ -1,7 +1,6 @@
 import { BasicAppContext } from '@mocks/helpers/app'
-import { fireEvent, render } from '@testing-library/react-native'
+import { render } from '@testing-library/react-native'
 import React from 'react'
-import { StyleSheet } from 'react-native'
 import TransferQRScannerScreen from './TransferQRScannerScreen'
 
 const mockNavigation = {
@@ -44,18 +43,5 @@ describe('TransferQRScanner', () => {
     )
 
     expect(tree).toMatchSnapshot()
-  })
-
-  it('shrinks the instruction text to its widest rendered line so it centers with the reticle', () => {
-    const { getByText } = render(
-      <BasicAppContext>
-        <TransferQRScannerScreen navigation={mockNavigation} />
-      </BasicAppContext>
-    )
-
-    const message = getByText('BCSC.Scan.WillScanAutomatically')
-    fireEvent(message, 'textLayout', { nativeEvent: { lines: [{ width: 156.5 }, { width: 120 }] } })
-
-    expect(StyleSheet.flatten(message.props.style).width).toBe(157)
   })
 })
